moved LangDropdown to shared
