+++ src/components/TestExampleLangDropdown/TestExampleLangDropdown.jsx
@@ -1,5 +1,5 @@
 import styles from "./TestExampleLangDropdown.module.scss";
-import LangDropdown from "../LangDropdown/LangDropdown";
+import LangDropdown from "../shared/LangDropdown/LangDropdown";
 import Logo from "../shared/Logo/Logo";
 
 export default function TestExampleLangDropdown() {
