+++ src/components/shared/LangDropdown/LangDropdown.module.scss
@@ -1,5 +1,5 @@
-@import "../../styles/variables";
-@import "../../styles/mixins";
+@import "../../../styles/variables";
+@import "../../../styles/mixins";
 
 .box {
   color: $secondary-white;
